+++ tgbotapi.extensions.utils/src/commonMain/kotlin/dev/inmo/tgbotapi/extensions/utils/SlotMachineUtils.kt
@@ -6,16 +6,30 @@ import dev.inmo.tgbotapi.types.dice.SlotMachineDiceAnimationType
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.Transient
 
-enum class SlotMachineReelImages {
-    BAR, BERRIES, LEMON, SEVEN
+/**
+ * @param text Is a text representation
+ * @param number Internal representation of reel
+ */
+enum class SlotMachineReelImage(val text: String, val number: Int) {
+    BAR("[bar]", 0),
+    BERRIES("\uD83C\uDF52", 1),
+    LEMON("\uD83C\uDF4B", 2),
+    SEVEN("7", 3)
 }
+@Deprecated("Renamed", ReplaceWith("SlotMachineReelImage", "dev.inmo.tgbotapi.extensions.utils.SlotMachineReelImage"))
+typealias SlotMachineReelImages = SlotMachineReelImage
+
+/**
+ * @return First [SlotMachineReelImage] with [SlotMachineReelImage.number] equal to receiver OR [SlotMachineReelImage.SEVEN]
+ */
 val Int.asSlotMachineReelImage
-    get() = when (this) {
-        0 -> SlotMachineReelImages.BAR
-        1 -> SlotMachineReelImages.BERRIES
-        2 -> SlotMachineReelImages.LEMON
-        else -> SlotMachineReelImages.SEVEN
-    }
+    get() = SlotMachineReelImage.values().firstOrNull { it.number == this } ?: SlotMachineReelImage.SEVEN
+
+/**
+ * @return First [SlotMachineReelImage] with [SlotMachineReelImage.text] equal to receiver OR [SlotMachineReelImage.SEVEN]
+ */
+val String.asSlotMachineReelImage
+    get() = SlotMachineReelImage.values().firstOrNull { it.text == this } ?: SlotMachineReelImage.SEVEN
 
 @Serializable
 data class SlotMachineResult(