+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 * `Common`:
     * `Version`:
         * `MicroUtils`: `0.4.24` -> `0.4.25`
+* `Extensions Utils`:
+    * `SlotMachineReelImages` has been renamed to `SlotMachineReelImage`
+    * `SlotMachineReelImage` got two built-in parameters: `text` and `number`
+    * New extension `String.asSlotMachineReelImage`
 
 ## 0.32.5